+++ CHANGELOG.md
@@ -80,6 +80,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - fix: .gitattributes stats, update langs case by @koru (#257)
 - Fix file rename on FUSE-based filesystems (#269) by @mxw
 - Fix history navigation in multibuffers (#275)
+- Fix incorrect window size on X11 (#259)
 
 ### Changed